중복 로직 분리
